Fix a few conversion warnings

+++ src/object/pneumatic_platform.cpp
@@ -96,7 +96,7 @@ PneumaticPlatform::update(float elapsed_time)
     movement = Vector(0, (start_y + offset_y) - get_pos().y);
   }
   if (this == master) {
-    int contact_diff = contacts.size() - slave->contacts.size();
+    size_t contact_diff = contacts.size() - slave->contacts.size();
     contacts.clear();
     slave->contacts.clear();
